requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0
aiofiles>=23.1.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
asyncpg>=0.27.0
psycopg2-binary>=2.9.5
aiosqlite>=0.19.0
google-cloud-storage>=2.8.0
google-cloud-firestore>=2.10.0
google-cloud-logging>=3.5.0
google-cloud-secret-manager>=2.16.0
pillow>=9.5.0
tenacity>=8.2.2
gunicorn>=20.1.0
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
mypy>=1.2.0
python-dotenv>=1.0.0
aiohttp>=3.8.5
playwright>=1.41.0